pyproject.toml
[tool.poetry]
name = "alzheimers-risk-app"
version = "0.1.0"
description = "Alzheimer's Risk Assessment Application"
authors = ["ABC <your.email@example.com>"]
readme = "README.md"
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.11.0,<3.12"
openai = "^1.3"
flask = "^3.1.1"
flask-session = "^0.8.0"
crewai = "^0.140.0"
numpy = "^2.3.1"
matplotlib = "^3.10.3"
reportlab = "^4.4.2"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
gunicorn = "^22.0.0"

[tool.poetry.dev-dependencies]
debugpy = "^1.6.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"